+++ network.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <math.h>
 
-float sigmoid(float z);
-float sigmoid_prime(float z);
+double sigmoid(float z);
+double sigmoid_prime(float z);
 void init_bias_and_weight();
